chore: consolidate api routes
