--- api/diag/health.js
@@ -1,3 +1,0 @@
-export default async function(req, res) {
-  res.status(200).json({ ok: true, time: new Date().toISOString() });
-}
