+++ vercel.json
@@ -1,6 +1,20 @@
 {
   "rewrites": [
-    { "source": "/check", "destination": "/check.html" }
+    { "source": "/check", "destination": "/check.html" },
+    { "source": "/api/diag/health",      "destination": "/api/router?action=diag_health" },
+    { "source": "/api/diag/notion",      "destination": "/api/router?action=diag_notion" },
+    { "source": "/api/diag/stripe",      "destination": "/api/router?action=diag_stripe" },
+    { "source": "/api/diag/telegram",    "destination": "/api/router?action=diag_telegram" },
+    { "source": "/api/diag/drive",       "destination": "/api/router?action=diag_drive" },
+    { "source": "/api/diag/cron",        "destination": "/api/router?action=diag_cron" },
+    { "source": "/api/stripe/audit",     "destination": "/api/router?action=stripe_audit" },
+    { "source": "/api/stripe/webhook",   "destination": "/api/router?action=stripe_webhook" },
+    { "source": "/api/gmail/scan",       "destination": "/api/router?action=gmail_scan" },
+    { "source": "/api/gmail/nudge",      "destination": "/api/router?action=gmail_nudge" },
+    { "source": "/api/cron/daily",       "destination": "/api/router?action=cron_daily" },
+    { "source": "/api/cron/digest",      "destination": "/api/router?action=cron_digest" },
+    { "source": "/api/cron/stripe-poll", "destination": "/api/router?action=cron_stripe_poll" },
+    { "source": "/api/cron/gmail-scan",  "destination": "/api/router?action=cron_gmail_scan" }
   ],
   "headers": [
     {
